chore: allow Owner to add rewards instead of Manager (#83)
### Description

Allow Owner to add rewards instead of Manager.

We assume the Manager will be in control of the overall funds (can
deposit and withdraw), but they won't assign individual rewards on a
regular basis.

We expect this will be done manually by our team for now. So we reserve
this function for the contract owner.

### Test plan

* Updated unit tests

### Related issues

* Related to RET-204